+++ CMakeLists.txt
@@ -42,7 +42,7 @@ endif()
 
 include(${LIBPARAMS_PATH}/CMakeLists.txt)
 add_definitions(-DFLASH_DRIVER_STORAGE_FILE="${APPLICATION_DIR}/default_params.yml")
-list(APPEND APPLICATION_SOURCES ${APPLICATION_DIR}/modules/modules.cpp)
+list(APPEND APPLICATION_SOURCES ${APPLICATION_DIR}/modules.cpp)
 list(APPEND APPLICATION_HEADERS ${ROOT_DIR}/Src ${APPLICATION_DIR})
 include(${APPLICATION_DIR}/CMakeLists.txt)
