+++ Src/cyphal_application/CMakeLists.txt
@@ -2,8 +2,9 @@
 # Distributed under the terms of the GPL v3 license, available in the file LICENSE.
 
 set(APPLICATION_DIR ${CMAKE_CURRENT_LIST_DIR})
+cmake_path(GET CMAKE_CURRENT_LIST_DIR PARENT_PATH SRC_DIR)
 
-include(${APPLICATION_DIR}/modules/cyphal/CMakeLists.txt)
+include(${SRC_DIR}/modules/cyphal_module/CMakeLists.txt)
 include(${APPLICATION_DIR}/modules/circuit_status/CMakeLists.txt)
 include(${APPLICATION_DIR}/modules/feedback/CMakeLists.txt)
 include(${APPLICATION_DIR}/modules/setpoint/CMakeLists.txt)
