+++ Src/dronecan_application/CMakeLists.txt
@@ -2,7 +2,8 @@
 # Distributed under the terms of the GPL v3 license, available in the file LICENSE.
 
 set(APPLICATION_DIR ${CMAKE_CURRENT_LIST_DIR})
+cmake_path(GET CMAKE_CURRENT_LIST_DIR PARENT_PATH SRC_DIR)
 
-include(${APPLICATION_DIR}/modules/dronecan/CMakeLists.txt)
+include(${SRC_DIR}/modules/dronecan_module/CMakeLists.txt)
 include(${APPLICATION_DIR}/modules/circuit_status/CMakeLists.txt)
 include(${APPLICATION_DIR}/modules/pwm/CMakeLists.txt)
